Bump cli-color to 1.1.0 and make sure to clear formatting on rest of line
The test summary line is output with a background color. On Fedora Linux with
the standard Gnome terminal, the background color is erroneously carried over
to the next line (which has the command prompt).

This commit clears the formatting on the rest of the line after outputting.
For this purpose, the cli-color dependency had to be updated to the most
recent version.